#1319 rename view
